+++ build_runner/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Mention --build-filter option in the README.
 - Update to build_daemon 4.0.
+- Support enabling experiments in build scripts.
 
 ## 2.4.0